% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createFitOutcomeModelArgs}
\alias{createFitOutcomeModelArgs}
\title{Create a parameter object for the function fitOutcomeModel}
\usage{
createFitOutcomeModelArgs(
  modelType = "logistic",
  stratified = FALSE,
  useCovariates = FALSE,
  inversePtWeighting = FALSE,
  bootstrapCi = FALSE,
  bootstrapRegularization = FALSE,
  interactionCovariateIds = c(),
  excludeCovariateIds = c(),
  includeCovariateIds = c(),
  profileGrid = NULL,
  profileBounds = c(log(0.1), log(10)),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", seed = 1, resetCoefficients = TRUE,
    startingVariance = 0.01, tolerance = 2e-07, cvRepetitions = 10, noiseLevel = "quiet")
)
}
\arguments{
\item{modelType}{The type of outcome model that will be used. Possible values are "logistic", "poisson", or "cox".}

\item{stratified}{Should the regression be conditioned on the strata defined in the population object (e.g. by matching or stratifying on propensity scores)?}

\item{useCovariates}{Whether to use the covariates in the cohortMethodData object in the outcome model.}

\item{inversePtWeighting}{Use inverse probability of treatment weighting (IPTW)}

\item{bootstrapCi}{Compute confidence interval using bootstrapping instead of likelihood profiling?}

\item{interactionCovariateIds}{An optional vector of covariate IDs to use to estimate interactions with the main treatment effect.}

\item{excludeCovariateIds}{Exclude these covariates from the outcome model.}

\item{includeCovariateIds}{Include only these covariates in the outcome model.}

\item{profileGrid}{A one-dimensional grid of points on the log(relative risk) scale where the likelihood for coefficient of variables is sampled. See details.}

\item{profileBounds}{The bounds (on the log relative risk scale) for the adaptive sampling of the likelihood function. See details.}

\item{prior}{The prior used to fit the model. See Cyclops::createPrior() for details.}

\item{control}{The control object used to control the cross-validation used to determine the hyperparameters of the prior (if applicable). See Cyclops::createControl() for details.}
}
\description{
Create a parameter object for the function fitOutcomeModel
}
\details{
Create an object defining the parameter values.
}
